+++ CHANGELOG.md
@@ -959,3 +959,10 @@ Version 0.4.1 (03-03-2025)
 --------------------------
 
 * Add combine operation for TSL keys TSL values -> TSD
+
+
+Version 0.4.2 (04-03-2025)
+--------------------------
+
+* Add combine operator for TS[tuple] and TS[tuple] -> TSD
+* Add values_ operator for TSD[K, TS[V]] -> TSS[V]